Reorganise icons
