+++ app/pages/lab-pages-editor/icons/assets/close-icon.png
