+++ app/pages/lab-pages-editor/icons/assets/options-icon.png
